+++ src/main/java/com/cg/cropbiddinginsuranceapp/service/IBankService.java
@@ -13,11 +13,11 @@ public interface IBankService {
 
 	public Bank addBank(Bank bank);
 
-	public Bank updateBank(long id, Bank bank);
+	public Bank updateBank(int id, Bank bank);
 
-	public Bank deleteBankById(long id);
+	public Bank deleteBankById(int id);
 
 	public List<Bank> retrieveAllBanks();
 
-	public Bank retrieveBankById(long id);
+	public Bank retrieveBankById(int id);
 }
